Separated components from New Project page
